requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
atari-py==0.2.9
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==1.2.2
ConfigArgParse==1.5.3
configparser==5.2.0
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.5.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
flatbuffers==1.12
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
gym==0.15.7
h5py==3.7.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.3.2
jupyter-core==4.10.0
jupyter-packaging==0.12.1
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kornia==0.6.5
libclang==14.0.1
lockfile==0.12.2
lpips==0.1.4
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mjcf2urdf==0.0.1
mpi4py==3.0.3
multidict==6.0.2
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.3
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.2
oauthlib==3.2.0
open3d==0.15.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plyfile==0.7.4
progressbar2==4.0.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser==2.21
pyDeprecate==0.3.0
pyglet==1.5.0
Pygments==2.12.0
pyparsing==3.0.7
pyquaternion==0.9.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-utils==3.3.3
pytorch-lightning==1.3.5
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==23.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.5.4
Send2Trash==1.8.0
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
subprocess32==3.5.4
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
test-tube==0.7.5
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tomlkit==0.11.0
torch==1.11.0
torchmetrics==0.9.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.10.33
warmup-scheduler==0.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
xlrd==1.2.0
yarl==1.7.2
zipp==3.8.0
